Implemented filters. Started with links. Have some problems with group destruction. This keeps the file open and leads to problems in the Group unit test. Need to fix this.
Changed the return values of the create<> Methods of the NXGroup class to first class objects. Otherwise we have problems with the return pointers - in particular we do not save anything. However, some testing is still required. Some of the constructors for the Group object are called at times where the class is no longer active.