Added garbage collector.
Update issue 54
State: WontFix
Labels: Priority-Low
This bug is related to the HDf5 library. We cannot fix this thus the status of
the bug has been reset to WontFix. As the bug becomes apparent only in cases
where the uses makes tremendously stupid choices in the chunk size the priority
will be reduced to Low.