Fixed an issue with linking
When the target path was relative and started with .. or a sequence of  ..
the link failed as HDF5 does not support full Unix semantics on its path 
objects. This issue is now fixed. However it should be mentioned in the 
manual that not full UNIX semantics is supported on the path objects.